--- PasteleriaAplicacion/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=54b01c1e
-build.xml.script.CRC32=1129032b
-build.xml.stylesheet.CRC32=8064a381@1.80.1.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=09787c8b
-nbproject/build-impl.xml.script.CRC32=9fc8a88f
-nbproject/build-impl.xml.stylesheet.CRC32=830a3534@1.80.1.48
