--- PasteleriaAplicacion/nbproject/private/configs/Run_in_Browser.properties
@@ -1,4 +1,0 @@
-# Do not modify this property in this configuration. It can be re-generated.
-javafx.run.as=embedded
-javafx.run.inbrowser=<Default System Browser>
-javafx.run.inbrowser.path=C:\\Program Files\\Internet Explorer\\iexplore.exe
