--- PasteleriaAplicacion/nbproject/private/private.properties
@@ -1,2 +1,0 @@
-auxiliary.org-netbeans-modules-projectapi.issue214819_5f_fx_5f_enabled=true
-user.properties.file=C:\\Users\\leodz\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
