--- PasteleriaAplicacion/nbproject/private/private.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
-    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/Users/leodz/OneDrive/Documentos/UADY%20LIS/3%20semestre/Programación%20orientada%20a%20objetos/Proyecto%20final/app-pasteleria/PasteleriaAplicacion/src/controlador/ControladorVistaRegistrarCliente.java</file>
-            <file>file:/C:/Users/leodz/OneDrive/Documentos/UADY%20LIS/3%20semestre/Programación%20orientada%20a%20objetos/Proyecto%20final/app-pasteleria/PasteleriaAplicacion/src/controlador/ControladorCliente.java</file>
-            <file>file:/C:/Users/leodz/OneDrive/Documentos/UADY%20LIS/3%20semestre/Programación%20orientada%20a%20objetos/Proyecto%20final/app-pasteleria/PasteleriaAplicacion/src/controlador/Main.java</file>
-        </group>
-    </open-files>
-</project-private>
