--- PasteleriaAplicacion/src/vista/VistaRegistrarCliente.fxml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-
-<AnchorPane id="AnchorPane" prefHeight="554.0" prefWidth="888.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8" fx:controller="controlador.ControladorVistaRegistrarCliente">
-   <children>
-      <TableView fx:id="tablaClientes" layoutX="12.0" layoutY="14.0" onMouseClicked="#seleccionar" prefHeight="407.0" prefWidth="864.0">
-         <columns>
-            <TableColumn fx:id="colID" prefWidth="202.0" text="ID" />
-            <TableColumn fx:id="colNombre" minWidth="8.0" prefWidth="208.0" text="Nombre" />
-            <TableColumn fx:id="colApellido" prefWidth="222.0" text="Apellido" />
-            <TableColumn fx:id="colSweetpoints" prefWidth="231.0" text="Sweetpoints" />
-         </columns>
-      </TableView>
-      <Button fx:id="btnEliminar" layoutX="593.0" layoutY="501.0" mnemonicParsing="false" onAction="#eliminar" prefHeight="31.0" prefWidth="116.0" text="Eliminar" />
-      <Button fx:id="btnModificar" layoutX="396.0" layoutY="501.0" mnemonicParsing="false" onAction="#modificar" prefHeight="31.0" prefWidth="116.0" text="Modificar" />
-      <Button fx:id="btnAgregar" layoutX="208.0" layoutY="501.0" mnemonicParsing="false" onAction="#agregar" prefHeight="31.0" prefWidth="116.0" text="Agregar" />
-      <Label fx:id="Apellido" layoutX="565.0" layoutY="446.0" text="Apellido" />
-      <TextField fx:id="textApellido" layoutX="637.0" layoutY="441.0" />
-      <Label fx:id="Nombre" layoutX="269.0" layoutY="446.0" text="Nombre" />
-      <TextField fx:id="textNombre" layoutX="340.0" layoutY="441.0" />
-      <TextField fx:id="textID" layoutX="67.0" layoutY="441.0" prefHeight="31.0" prefWidth="149.0" />
-      <Label fx:id="ID" layoutX="32.0" layoutY="446.0" text="ID" />
-   </children>
-</AnchorPane>
