+++ PasteleriaApp/nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=2e13ddbe
+build.xml.script.CRC32=a7d42d60
+build.xml.stylesheet.CRC32=8064a381@1.80.1.48
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=2e13ddbe
+nbproject/build-impl.xml.script.CRC32=8558fa45
+nbproject/build-impl.xml.stylesheet.CRC32=830a3534@1.80.1.48
