+++ PasteleriaApp/nbproject/private/configs/Run_in_Browser.properties
@@ -1,0 +1,2 @@
+# Do not modify this property in this configuration. It can be re-generated.
+javafx.run.as=embedded
