+++ PasteleriaApp/nbproject/private/private.properties
@@ -1,0 +1,6 @@
+auxiliary.org-netbeans-modules-projectapi.issue214819_5f_fx_5f_enabled=true
+# No need to modify this property unless customizing JavaFX Ant task infrastructure
+endorsed.javafx.ant.classpath=.
+javafx.run.inbrowser=<Default System Browser>
+javafx.run.inbrowser.path=C:\\Program Files\\Internet Explorer\\IEXPLORE.EXE
+user.properties.file=C:\\Users\\johnc\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
